ci: fix unnecessary conversion

+++ pkg/security/secl/model/unmarshallers.go
@@ -221,14 +221,14 @@ func (e *ExitEvent) UnmarshalBinary(data []byte) (int, error) {
 	exitStatus := ByteOrder.Uint32(data[0:4])
 	if exitStatus&0x7F == 0x00 { // process terminated normally
 		e.Cause = uint32(ExitExited)
-		e.Code = uint32(exitStatus>>8) & 0xFF
+		e.Code = (exitStatus >> 8) & 0xFF
 	} else if exitStatus&0x7F != 0x7F { // process terminated because of a signal
 		if exitStatus&0x80 == 0x80 { // coredump signal
 			e.Cause = uint32(ExitCoreDumped)
-			e.Code = uint32(exitStatus & 0x7F)
+			e.Code = exitStatus & 0x7F
 		} else { // other signals
 			e.Cause = uint32(ExitSignaled)
-			e.Code = uint32(exitStatus & 0x7F)
+			e.Code = exitStatus & 0x7F
 		}
 	}
